Add tests for arrow=37 to arrow=49
